Add files via upload
The vacuum cleaner problem is a goal-based agent problem in artificial intelligence (AI). The goal is to clean an area by using a vacuum cleaner. The vacuum cleaner has a location sensor and a dirt sensor to know where it is and if the room is dirty. The vacuum cleaner can move left, right, suck, and idle.