Add files via upload
We assign a score to the possible moves in the game and evaluate the best possible move that maximizes the computer's chance to win. One player is assigned to be the “Max player” who try to maximize his score and the other player is the “Min player”, minimizing his score.